app/requirements.txt
flask
python-dotenv
pytz
requests
prometheus_client
